refactor: コンポーネントの整理
