+++ src/components/ComponentStory/index.module.css
