+++ src/components/article/Article.astro
@@ -1,5 +1,7 @@
 ---
 import styles from './Article.module.css';
+// NOTE: as はこの Issue の回避のため
+// https://github.com/withastro/language-tools/issues/580
 import { Article as UIArticle } from 'smarthr-ui';
 
 type Props = {
