--- src/components/article/CodeBlock/index.ts
@@ -1,2 +1,0 @@
-export { CodeBlock, type LiveContainerProps } from './CodeBlock';
-export { DesignPatternCodeBlock } from './DesignPatternCodeBlock';
