+++ src/pages/[...slug].astro
@@ -2,7 +2,7 @@
 import { getCollection } from 'astro:content';
 import ArticleLayout from '@/layouts/ArticleLayout.astro';
 import { CustomLink } from '@/components/article/CustomLink';
-import CodeBlock from '@/components/MDXComponents/CodeBlock.astro';
+import CodeBlock from '@/components/article/CodeBlock.astro';
 import { FragmentTitle } from '@/components/article/FragmentTitle';
 
 export async function getStaticPaths() {
